app/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apache-airflow==2.7.2
pandas==2.0.3
numpy==1.24.3